tidy TripPlanner component
